refactor break into multiple components
